Cleaned up and renamed files
