--- false-positives.json
@@ -1,1 +1,0 @@
-{"VIX": "'regularMarketOpen'", "LLC": "'regularMarketOpen'", "BAN": "'regularMarketOpen'", "YOLO": "index 0 is out of bounds for axis 0 with size 0", "THE": "'regularMarketOpen'", "POWER": "'regularMarketOpen'", "MUSK": "'regularMarketOpen'", "YOU": "'regularMarketOpen'", "WSB": "'regularMarketOpen'", "WHO": "'regularMarketOpen'", "LOST": "'regularMarketOpen'", "MONEY": "'regularMarketOpen'", "SLVR": "'regularMarketOpen'", "CALLS": "'regularMarketOpen'", "SEC": "'regularMarketOpen'", "EOM": "'regularMarketOpen'", "BAE": "'regularMarketOpen'", "LEAP": "'regularMarketOpen'", "ONLY": "'regularMarketOpen'", "AND": "'regularMarketOpen'", "CNBC": "'regularMarketOpen'", "GOP": "index 0 is out of bounds for axis 0 with size 0", "CRASH": "'regularMarketOpen'", "SHORT": "'regularMarketOpen'", "THIS": "'regularMarketOpen'", "FAULT": "'regularMarketOpen'", "THAT": "'regularMarketOpen'", "TSMC": "'regularMarketOpen'", "POOTS": "BeautifulSoup4 (bs4) not found, please install it", "MOON": "'regularMarketOpen'", "MARS": "'regularMarketOpen'", "LETS": "'regularMarketOpen'", "AFTER": "'regularMarketOpen'", "HOUR": "'regularMarketOpen'", "EOY": "'regularMarketOpen'", "BEZOS": "BeautifulSoup4 (bs4) not found, please install it", "FUCK": "'regularMarketOpen'", "YOUR": "'regularMarketOpen'", "WIVES": "'regularMarketOpen'", "PEACE": "'regularMarketOpen'", "SHITS": "'regularMarketOpen'", "WTF": "'regularMarketOpen'", "RIGHT": "'regularMarketOpen'", "FELT": "'regularMarketOpen'", "INTEL": "'regularMarketOpen'", "OTM": "'regularMarketOpen'", "SEES": "'regularMarketOpen'", "ARM": "'regularMarketOpen'", "APPL": "'regularMarketOpen'", "LFG": "'regularMarketOpen'", "GUH": "'regularMarketOpen'", "INTL": "BeautifulSoup4 (bs4) not found, please install it", "CALL": "'regularMarketOpen'", "JPOW": "'regularMarketOpen'", "GOING": "'regularMarketOpen'", "UNDER": "'regularMarketOpen'", "BLANK": "'regularMarketOpen'", "CHECK": "'regularMarketOpen'", "GET": "'regularMarketOpen'", "SAVED": "'regularMarketOpen'", "LMAO": "'regularMarketOpen'", "DAILY": "'regularMarketOpen'", "RUG": "'regularMarketOpen'", "PULL": "'regularMarketOpen'", "GREEN": "'regularMarketOpen'", "RED": "'regularMarketOpen'", "FFS": "'regularMarketOpen'", "CPU": "'regularMarketOpen'", "TENDY": "'regularMarketOpen'", "COVID": "'regularMarketOpen'", "RIP": "'regularMarketOpen'", "BULLS": "'regularMarketOpen'", "BEND": "'regularMarketOpen'", "FUK": "'regularMarketOpen'", "OVER": "'regularMarketOpen'", "CAUSE": "'regularMarketOpen'", "BEARS": "'regularMarketOpen'", "HERE": "'regularMarketOpen'", "FEAST": "'regularMarketOpen'", "ETF": "'regularMarketOpen'", "IDK": "'regularMarketOpen'", "NBC": "'regularMarketOpen'", "OMG": "'regularMarketOpen'", "FAGMA": "'regularMarketOpen'", "LARP": "'regularMarketOpen'", "TLDR": "'regularMarketOpen'", "CANT": "'regularMarketOpen'", "LOSE": "'regularMarketOpen'", "SURE": "'regularMarketOpen'", "LOL": "'regularMarketOpen'", "BTNX": "'regularMarketOpen'", "DDTG": "'regularMarketOpen'", "NYC": "'regularMarketOpen'", "RUGER": "'regularMarketOpen'", "BDSM": "'regularMarketOpen'", "DAU": "'regularMarketOpen'", "FOMO": "'regularMarketOpen'", "SVB": "'regularMarketOpen'", "HVAC": "'regularMarketOpen'", "YEARS": "'regularMarketOpen'", "RISK": "BeautifulSoup4 (bs4) not found, please install it", "GTFO": "'regularMarketOpen'", "CNN": "'regularMarketOpen'", "MSNBC": "'regularMarketOpen'", "UNITE": "'regularMarketOpen'", "LEAPS": "'regularMarketOpen'", "BALLS": "'regularMarketOpen'", "HAVE": "'regularMarketOpen'", "DICK": "'regularMarketOpen'", "BTW": "'regularMarketOpen'", "MUTED": "'regularMarketOpen'", "HGM": "'regularMarketOpen'", "PILES": "'regularMarketOpen'", "FED": "'regularMarketOpen'", "JUL": "'regularMarketOpen'", "RSI": "'regularMarketOpen'", "MACD": "'regularMarketOpen'", "SPX": "'regularMarketOpen'", "SLA": "'regularMarketOpen'", "GAVE": "'regularMarketOpen'", "THROW": "'regularMarketOpen'", "SWEAR": "'regularMarketOpen'", "WRONG": "BeautifulSoup4 (bs4) not found, please install it", "PISS": "'regularMarketOpen'", "OFF": "'longName'", "HOLY": "'regularMarketOpen'", "WHAT": "'regularMarketOpen'", "WERE": "'regularMarketOpen'", "LOOK": "'regularMarketOpen'", "POTUS": "BeautifulSoup4 (bs4) not found, please install it", "PPP": "'regularMarketOpen'", "HIGH": "'regularMarketOpen'", "OCT": "'regularMarketOpen'", "LUL": "'regularMarketOpen'", "SMA": "'regularMarketOpen'", "NEVER": "'regularMarketOpen'", "IMHO": "'regularMarketOpen'", "DXY": "'regularMarketOpen'", "POV": "'regularMarketOpen'", "NYSE": "'regularMarketOpen'", "ADHD": "'regularMarketOpen'", "TODAY": "'regularMarketOpen'", "SUCKS": "BeautifulSoup4 (bs4) not found, please install it", "AMA": "'regularMarketOpen'", "TFW": "'regularMarketOpen'", "FBI": "'regularMarketOpen'", "TARP": "'regularMarketOpen'", "STILL": "'regularMarketOpen'", "NOT": "'regularMarketOpen'", "UGLY": "'regularMarketOpen'", "XBOX": "'regularMarketOpen'", "COD": "'regularMarketOpen'", "YOY": "'regularMarketOpen'", "GULAG": "'regularMarketOpen'", "MAU": "'regularMarketOpen'", "TOS": "'regularMarketOpen'", "PUTS": "'regularMarketOpen'", "VVIX": "'regularMarketOpen'", "GEX": "index 0 is out of bounds for axis 0 with size 0", "DIX": "'regularMarketOpen'", "DID": "'regularMarketOpen'", "SAY": "'regularMarketOpen'", "HOF": "'regularMarketOpen'", "ALSO": "'regularMarketOpen'", "TIME": "'regularMarketOpen'", "IRA": "'regularMarketOpen'", "SIZES": "'regularMarketOpen'", "PAY": "'regularMarketOpen'", "FYI": "'regularMarketOpen'", "FAANG": "'regularMarketOpen'", "TIL": "BeautifulSoup4 (bs4) not found, please install it", "YES": "'regularMarketOpen'", "FOMC": "'regularMarketOpen'", "FREE": "'regularMarketOpen'", "XXXX": "BeautifulSoup4 (bs4) not found, please install it", "YYYY": "'regularMarketOpen'", "CRUSH": "'regularMarketOpen'", "YTD": "'regularMarketOpen'", "MYTH": "'regularMarketOpen'", "TDA": "BeautifulSoup4 (bs4) not found, please install it", "ADR": "'regularMarketOpen'", "FDA": "'regularMarketOpen'", "FAR": "'regularMarketOpen'", "ATM": "'regularMarketOpen'", "GAP": "'regularMarketOpen'"}
